Player and difficulty selectors
Changed the scripts for the player number and difficulty level to just bool values.
Created public methods for the buttons to call when the buttons are clicked.
Added the script to the buttons.
Added a back button method for the players and difficulty so that it changes the bool values to false when the back button is clicked.